+++ lib/ex_bashkit/native.ex
@@ -39,11 +39,19 @@ defmodule ExBashkit.Native do
         _mounts,
         _allowed_mount_paths,
         _limits,
-        _network
+        _network,
+        _builtin_names
       ),
       do: :erlang.nif_error(:nif_not_loaded)
 
-  def session_exec(_session, _script), do: :erlang.nif_error(:nif_not_loaded)
+  def session_exec(_session, _script, _handler, _builtin_timeout_ms),
+    do: :erlang.nif_error(:nif_not_loaded)
+
   def session_read_file(_session, _path), do: :erlang.nif_error(:nif_not_loaded)
   def session_write_file(_session, _path, _content), do: :erlang.nif_error(:nif_not_loaded)
+
+  # Delivers an Elixir builtin handler's result back into the waiting Rust
+  # builtin (see ExBashkit.Session back-call handler).
+  def builtin_reply(_req_id, _stdout, _stderr, _exit_code),
+    do: :erlang.nif_error(:nif_not_loaded)
 end